Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/character_solver/collision_error_function.cpp
@@ -248,7 +248,7 @@ std::vector<Vector2i> CollisionErrorFunctionT<T>::getCollisionPairs() const {
     Vector2<T> cp;
     T overlap;
     if (checkCollision(collisionState_, pair.indexA, pair.indexB, distance, cp, overlap)) {
-      collidingPairs.emplace_back(pair.indexA, pair.indexB);
+      collidingPairs.emplace_back(static_cast<int>(pair.indexA), static_cast<int>(pair.indexB));
     }
   }
   return collidingPairs;
